Addressed comments 2
