+++ flink-formats/flink-json/src/main/java/org/apache/flink/formats/json/JsonRowDeserializationSchema.java
@@ -20,17 +20,44 @@ package org.apache.flink.formats.json;
 
 import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
+import org.apache.flink.api.common.typeinfo.BasicArrayTypeInfo;
+import org.apache.flink.api.common.typeinfo.PrimitiveArrayTypeInfo;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
+import org.apache.flink.api.common.typeinfo.Types;
+import org.apache.flink.api.java.typeutils.ObjectArrayTypeInfo;
 import org.apache.flink.api.java.typeutils.RowTypeInfo;
-import org.apache.flink.formats.json.RuntimeDeserializationConverterFactory.DeserializationRuntimeConverter;
 import org.apache.flink.types.Row;
+import org.apache.flink.util.WrappingRuntimeException;
 
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.JsonNode;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.node.ArrayNode;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.node.ObjectNode;
 
 import java.io.IOException;
+import java.io.Serializable;
+import java.sql.Date;
+import java.sql.Time;
+import java.sql.Timestamp;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.ZoneOffset;
+import java.time.temporal.TemporalAccessor;
+import java.time.temporal.TemporalQueries;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
+import static java.lang.String.format;
+import static java.time.format.DateTimeFormatter.ISO_LOCAL_DATE;
+import static java.util.Spliterators.spliterator;
+import static java.util.stream.StreamSupport.stream;
+import static org.apache.flink.formats.json.TimeFormats.RFC3339_TIMESTAMP_FORMAT;
+import static org.apache.flink.formats.json.TimeFormats.RFC3339_TIME_FORMAT;
 import static org.apache.flink.util.Preconditions.checkArgument;
 import static org.apache.flink.util.Preconditions.checkNotNull;
 
@@ -50,6 +77,8 @@ public class JsonRowDeserializationSchema implements DeserializationSchema<Row> 
 	/** Type information describing the result type. */
 	private final RowTypeInfo typeInfo;
 
+	private final boolean failOnMissingField;
+
 	/** Object mapper for parsing the JSON. */
 	private final ObjectMapper objectMapper = new ObjectMapper();
 
@@ -57,12 +86,12 @@ public class JsonRowDeserializationSchema implements DeserializationSchema<Row> 
 
 	private JsonRowDeserializationSchema(
 			TypeInformation<Row> typeInfo,
-			DeserializationRuntimeConverter runtimeConverter) {
+			boolean failOnMissingField) {
 		checkNotNull(typeInfo, "Type information");
 		checkArgument(typeInfo instanceof RowTypeInfo, "Only RowTypeInfo is supported");
-		checkNotNull(runtimeConverter, "Did not provide runtime converter.");
 		this.typeInfo = (RowTypeInfo) typeInfo;
-		this.runtimeConverter = runtimeConverter;
+		this.failOnMissingField = failOnMissingField;
+		this.runtimeConverter = createConverter(this.typeInfo);
 	}
 
 	@Override
@@ -93,7 +122,6 @@ public class JsonRowDeserializationSchema implements DeserializationSchema<Row> 
 
 		private final RowTypeInfo typeInfo;
 		private boolean failOnMissingField = false;
-		private RuntimeDeserializationConverterFactory converterFactory;
 
 		/**
 		 * Creates a JSON deserialization schema for the given type information.
@@ -127,20 +155,8 @@ public class JsonRowDeserializationSchema implements DeserializationSchema<Row> 
 			return this;
 		}
 
-		/**
-		 * Configures schema to use runtime converter factory.
-		 */
-		public Builder useConverterFactory(RuntimeDeserializationConverterFactory converterFactory) {
-			this.converterFactory = converterFactory;
-			return this;
-		}
-
 		public JsonRowDeserializationSchema build() {
-			RuntimeDeserializationConverterFactory factory =
-				converterFactory != null ? converterFactory : new DefaultRuntimeDeserializationConverterFactory();
-
-			return new JsonRowDeserializationSchema(typeInfo,
-				factory.getDeserializationRuntimeConverter(typeInfo, failOnMissingField));
+			return new JsonRowDeserializationSchema(typeInfo, failOnMissingField);
 		}
 	}
 
@@ -154,11 +170,216 @@ public class JsonRowDeserializationSchema implements DeserializationSchema<Row> 
 		}
 		final JsonRowDeserializationSchema that = (JsonRowDeserializationSchema) o;
 		return Objects.equals(typeInfo, that.typeInfo) &&
-			Objects.equals(runtimeConverter, that.runtimeConverter);
+			Objects.equals(failOnMissingField, that.failOnMissingField);
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(typeInfo, runtimeConverter);
+		return Objects.hash(typeInfo, failOnMissingField);
+	}
+
+	/*
+		Runtime converter
+	 */
+
+	/**
+	 * Runtime converter that maps between {@link JsonNode}s and Java objects.
+	 */
+	@FunctionalInterface
+	private interface DeserializationRuntimeConverter extends Serializable {
+		Object convert(ObjectMapper mapper, JsonNode jsonNode);
+	}
+
+	private DeserializationRuntimeConverter createConverter(TypeInformation<?> typeInfo) {
+		DeserializationRuntimeConverter baseConverter = createConverterForSimpleType(typeInfo)
+			.orElseGet(() ->
+				createContainerConverter(typeInfo)
+					.orElseGet(() -> createFallbackConverter(typeInfo.getTypeClass())));
+		return wrapIntoNullableConverter(baseConverter);
+	}
+
+	private DeserializationRuntimeConverter wrapIntoNullableConverter(DeserializationRuntimeConverter converter) {
+		return (mapper, jsonNode) -> {
+			if (jsonNode.isNull()) {
+				return null;
+			}
+
+			return converter.convert(mapper, jsonNode);
+		};
+	}
+
+	private Optional<DeserializationRuntimeConverter> createContainerConverter(TypeInformation<?> typeInfo) {
+		if (typeInfo instanceof RowTypeInfo) {
+			return Optional.of(createRowConverter((RowTypeInfo) typeInfo));
+		} else if (typeInfo instanceof ObjectArrayTypeInfo) {
+			return Optional.of(createObjectArrayConverter(((ObjectArrayTypeInfo) typeInfo).getComponentInfo()));
+		} else if (typeInfo instanceof BasicArrayTypeInfo) {
+			return Optional.of(createObjectArrayConverter(((BasicArrayTypeInfo) typeInfo).getComponentInfo()));
+		} else if (isPrimitiveByteArray(typeInfo)) {
+			return Optional.of(createByteArrayConverter());
+		} else {
+			return Optional.empty();
+		}
+	}
+
+	private DeserializationRuntimeConverter createByteArrayConverter() {
+		return (mapper, jsonNode) -> {
+			try {
+				return jsonNode.binaryValue();
+			} catch (IOException e) {
+				throw new WrappingRuntimeException("Unable to deserialize byte array.", e);
+			}
+		};
+	}
+
+	private boolean isPrimitiveByteArray(TypeInformation<?> typeInfo) {
+		return typeInfo instanceof PrimitiveArrayTypeInfo &&
+			((PrimitiveArrayTypeInfo) typeInfo).getComponentType() == Types.BYTE;
+	}
+
+	private DeserializationRuntimeConverter createObjectArrayConverter(TypeInformation elementTypeInfo) {
+		DeserializationRuntimeConverter elementConverter = createConverter(elementTypeInfo);
+		return assembleArrayConverter(elementConverter);
+	}
+
+	private DeserializationRuntimeConverter createRowConverter(RowTypeInfo typeInfo) {
+		List<DeserializationRuntimeConverter> fieldConverters = Arrays.stream(typeInfo.getFieldTypes())
+			.map(this::createConverter)
+			.collect(Collectors.toList());
+
+		return assembleRowConverter(typeInfo.getFieldNames(), fieldConverters);
+	}
+
+	private DeserializationRuntimeConverter createFallbackConverter(Class<?> valueType) {
+		return (mapper, jsonNode) -> {
+			// for types that were specified without JSON schema
+			// e.g. POJOs
+			try {
+				return mapper.treeToValue(jsonNode, valueType);
+			} catch (JsonProcessingException e) {
+				throw new WrappingRuntimeException(format("Could not convert node: %s", jsonNode), e);
+			}
+		};
+	}
+
+	private Optional<DeserializationRuntimeConverter> createConverterForSimpleType(TypeInformation<?> simpleTypeInfo) {
+		if (simpleTypeInfo == Types.VOID) {
+			return Optional.of((mapper, jsonNode) -> null);
+		} else if (simpleTypeInfo == Types.BOOLEAN) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.asBoolean());
+		} else if (simpleTypeInfo == Types.STRING) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.asText());
+		} else if (simpleTypeInfo == Types.INT) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.asInt());
+		} else if (simpleTypeInfo == Types.LONG) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.asLong());
+		} else if (simpleTypeInfo == Types.DOUBLE) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.asDouble());
+		} else if (simpleTypeInfo == Types.FLOAT) {
+			return Optional.of((mapper, jsonNode) -> Float.parseFloat(jsonNode.asText().trim()));
+		} else if (simpleTypeInfo == Types.SHORT) {
+			return Optional.of((mapper, jsonNode) -> Short.parseShort(jsonNode.asText().trim()));
+		} else if (simpleTypeInfo == Types.BYTE) {
+			return Optional.of((mapper, jsonNode) -> Byte.parseByte(jsonNode.asText().trim()));
+		} else if (simpleTypeInfo == Types.BIG_DEC) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.decimalValue());
+		} else if (simpleTypeInfo == Types.BIG_INT) {
+			return Optional.of((mapper, jsonNode) -> jsonNode.bigIntegerValue());
+		} else if (simpleTypeInfo == Types.SQL_DATE) {
+			return Optional.of(createDateConverter());
+		} else if (simpleTypeInfo == Types.SQL_TIME) {
+			return Optional.of(createTimeConverter());
+		} else if (simpleTypeInfo == Types.SQL_TIMESTAMP) {
+			return Optional.of(createTimestampConverter());
+		} else {
+			return Optional.empty();
+		}
+	}
+
+	private DeserializationRuntimeConverter createDateConverter() {
+		return (mapper, jsonNode) -> Date.valueOf(ISO_LOCAL_DATE.parse(jsonNode.asText())
+			.query(TemporalQueries.localDate()));
+	}
+
+	private DeserializationRuntimeConverter createTimestampConverter() {
+		return (mapper, jsonNode) -> {
+			TemporalAccessor parsedTimestamp = RFC3339_TIMESTAMP_FORMAT.parse(jsonNode.asText());
+
+			ZoneOffset zoneOffset = parsedTimestamp.query(TemporalQueries.offset());
+
+			if (zoneOffset != null && zoneOffset.getTotalSeconds() != 0) {
+				throw new IllegalStateException(
+					"Invalid timestamp format. Only a timestamp in UTC timezone is supported yet. " +
+						"Format: yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
+			}
+
+			LocalTime localTime = parsedTimestamp.query(TemporalQueries.localTime());
+			LocalDate localDate = parsedTimestamp.query(TemporalQueries.localDate());
+
+			return Timestamp.valueOf(LocalDateTime.of(localDate, localTime));
+		};
+	}
+
+	private DeserializationRuntimeConverter createTimeConverter() {
+		return (mapper, jsonNode) -> {
+
+			TemporalAccessor parsedTime = RFC3339_TIME_FORMAT.parse(jsonNode.asText());
+
+			ZoneOffset zoneOffset = parsedTime.query(TemporalQueries.offset());
+			LocalTime localTime = parsedTime.query(TemporalQueries.localTime());
+
+			if (zoneOffset != null && zoneOffset.getTotalSeconds() != 0 || localTime.getNano() != 0) {
+				throw new IllegalStateException(
+					"Invalid time format. Only a time in UTC timezone without milliseconds is supported yet.");
+			}
+
+			return Time.valueOf(localTime);
+		};
+	}
+
+	private DeserializationRuntimeConverter assembleRowConverter(
+		String[] fieldNames,
+		List<DeserializationRuntimeConverter> fieldConverters) {
+		return (mapper, jsonNode) -> {
+			ObjectNode node = (ObjectNode) jsonNode;
+
+			int arity = fieldNames.length;
+			Row row = new Row(arity);
+			for (int i = 0; i < arity; i++) {
+				String fieldName = fieldNames[i];
+				JsonNode field = node.get(fieldName);
+				Object convertField = convertField(mapper, fieldConverters.get(i), fieldName, field);
+				row.setField(i, convertField);
+			}
+
+			return row;
+		};
+	}
+
+	private Object convertField(
+		ObjectMapper mapper,
+		DeserializationRuntimeConverter fieldConverter,
+		String fieldName,
+		JsonNode field) {
+		if (field == null) {
+			if (failOnMissingField) {
+				throw new IllegalStateException(
+					"Could not find field with name '" + fieldName + "'.");
+			} else {
+				return null;
+			}
+		} else {
+			return fieldConverter.convert(mapper, field);
+		}
+	}
+
+	private DeserializationRuntimeConverter assembleArrayConverter(DeserializationRuntimeConverter elementConverter) {
+		return (mapper, jsonNode) -> {
+			ArrayNode node = (ArrayNode) jsonNode;
+
+			return stream(spliterator(node.elements(), node.size(), 0), false)
+				.map(innerNode -> elementConverter.convert(mapper, innerNode))
+				.toArray();
+		};
 	}
 }
